+++ lotus/views/mixins.py
@@ -91,3 +91,32 @@ class ArticleFilterMixin:
         context = super().get_context_data(**kwargs)
         context["lotus_now"] = self.target_date
         return context
+
+
+class LotusContextStage:
+    """
+    Mixin to inject Lotus stage into view context.
+
+    Lotus stage is commonly used for Lotus navigation, it just indicated where a view
+    is located from main Lotus content type views (Article, Author, Category, ..).
+
+    View which inherits from this mixin should set view attribute ``lotus_stage`` to
+    a main content type in lowercase like ``articles``, ``authors``, ``category``.
+
+    The default stage value is ``None``.
+
+    Finally, the Lotus stage is just an helper for basic navigation and there is no
+    code which use it so it can be ignored from custom Lotus implementation.
+    """
+    lotus_stage = None
+
+    def get_lotus_stage(self):
+        return self.lotus_stage
+
+    def get_context_data(self, **kwargs):
+        """
+        Expose the date "now" used for publication filter.
+        """
+        context = super().get_context_data(**kwargs)
+        context["lotus_stage"] = self.get_lotus_stage()
+        return context